Refactored for readability and utility
Refactored for readability and utility

Refactored with the goal of adhereing to principles put forward in
"Clean Code" by Robert C Martin. I'm reading it now and used this review as a chance to practice some of what I've read so far. Hopefully the end result is more readable and will be easier to maintain, change, or add features to the code.

I also updated the CLI such that it can provide a help message and take in an optional output file argument so you can specify which file to save to. I changed the way the package was structured so it can be installed with pip and then called from the command line. This way it can be run from anywhere on a computer and you don't have to think about where the package files live.

With these changes, the program takes a little under 4 seconds longer to run than the original version. Hopefully the added utility to the command line and the (hopefully) easier to maintain code make up for it.

Running the original version and this version side by side produce
almost the exact same output csv. The only difference is one small
formatting change I made which was adding an extra blank line between the plate data and the total core mass. Since this new version takes an optional third output file argument, the file names may be different as well.